Section 6 Part 3 - Blink Led With Timer Input Capture
Capturing input data from toggling LED PA5 -> PA6.  The program initializes data to the timestamp variable everytime the TIMER status register is set to high, it is set to high when the pin captures a rising edge in this case. Rising edge when LED turns on. From experiment, the toggle value increase by 2000 every led on event.