Project 1 - Rising Edge Counter
This project involves counting the number of button presses within a minute and displaying that number in RealTerm.
Utilizing the Timer2 to time a 60-sec interval, and the EXTI interrupt to count the number of times button pressed.
Note: The button press is not active due to debounce issue - on ocassion the interrupt may detect multiple events during one button press. To remedy this issue, an RC filter would need to be used.

The next part of this project would be to incorporate a keypad.